Return whole file if range request cannot be granted due to encryption
Whenenver range headers are set and encryption is enabled, it is not possible
to automatically fseek() to the proper position.

To avoid returning corrupt/invalid data or causing a decryption
error, the range headers are stripped so that the SabreDAV code in
httpGet() returns the whole file.

Backport of cc8c1d8e0776bda7f4c7c28e11984144a4861df3 from master