Be a bit more cautious when removing paths from the mapper
Spaces and special characters are replaced with - on windows.
So test@/ and test€/ have the same physical folder name test-/.
To not run into a problem, we append an index to them. So the
first folder is created as test-/, the second one as test--1/

Now if we delete test- (while deleting test@/) the old implementation
will also deletes test--1 which is test€/, so the next call will create
an entry mapping test€ to test- therefor we have 2 different map paths
for the same logical path, which breaks file finding, deleting, modifying, ...

Fix #6553